requirements.txt
azure-core==1.29.2
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Faker==19.3.0
idna==3.4
isodate==0.6.1
numpy==1.25.2
pandas==2.0.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
